Correctly use custom Stylus version if available
Per the README (https://github.com/gabrielgrant/broccoli-stylus-single#stylus-version) the Stylus version used by this plugin should be configurable.

`peerDependencies` doesn't appear to work in the way that I had expected it to.

This change will check for `stylus` in the root project and fall back to the version included as a dependency of `broccoli-stylus-single` if it's not present.